// src/app/core/models/masters/ms-audit-management/ms-audit-plan-criteria.ts
export interface MsAuditPlanCriteria {
    id: number;
    title: string;
    description: string;
    status: string;
    status_id: number;
    status_label: string;
}

export interface MsAuditPlanCriteriaPaginationResponse {
    current_page: number;
    from: number;
    total: number;
    per_page: number;
    last_page: number;
    data: MsAuditPlanCriteria[];
}